TRCReader --> Modified type for microseconds. It will be u64 instead of u32.
Added flowlayout.
Added StatusEdition widget

SPNs of type status can be now inspected.

+++ J1939Viewer/trcinspector.h
@@ -36,7 +36,7 @@ private:
     Ui::TRCInspector *ui;
     TRCLoader* mLoader;
 
-    QList< QPair< u32, J1939::J1939Frame* > > mFrameList;
+    QList< QPair< u64, J1939::J1939Frame* > > mFrameList;
     void startBackgroundLoader(const QString& file);
     void showNumericSPNViewer(NumericSPNItem*);
 };
